Update patch set 6
Patch Set 6: Code-Review-1

(7 comments)

This patch seems to have a number of problems and limitations.

Firstly, it's getting the config wrong - for instance, external and provider networks do not necessarily have the same MTU as internal networks (since they aren't L3-encapped) but I believe you've written this so that all networks have the same MTU.  (This implies in turn that router interfaces will have a variety of MTUs, even within one router, and DHCP servers shouldn't always advertise the same MTU.)

Secondly, when addressing L3 overlays, the segment MTU as found on an interface is not necessarily the MTU that limits tunnel packet sizes - that would be the path MTU, and could be limited by a hop anywhere between your IP and the IP of the machine you're reaching.  It's a possible-but-not-likely case, and you can override the value your code finds, but it's not at all clear to an end user that the code is *guessing* the MTU rather than reliably working it out; nothing in your patch's comments suggests an override *might* be required; and comments aside, a person running devstack would have absolutely no idea what was happening if this went wrong ('that's odd, some packets are disappearing but not others') and also no documentation to refer to short of reading all the devstack scripts.

Thirdly, advertising MTU via DHCPv4 is nice but insufficient.  You haven't fixed IPv6 RAs, which also advertise MTU.  I can run VMs that don't use DHCP or RA to configure their interfaces (where I configure them statically in a cloud-init script, for instance), and I wouldn't know that I had to set an MTU on them, because devstack does not at any point tell me.  Not sure what you can do about that particular point other than document it, but it hasn't been documented.

Finally, this is being addressed by a bunch of MTU selection and advertisement patches in Kilo so this patch may not have a long lifetime.

Your patch absolutely doesn't have to address every case in the world, it just has to improve on what happens at the moment.  But I think it does need to do a much better job of explaining the cases it fixes and the cases it doesn't, both to the code reviewer and to the end user.

Patch-set: 6
Reviewer: Gerrit User 3217 <3217@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1