Update patch set 3
Patch Set 3: Code-Review-1

this seems like a regression in that we're going from letting max_connections take the "upstream" default (151 at the moment I guess, but who knows what in the future) to hard-coding it ourselves.

IMO at least it should be in one place as a global in stackrc; and do nothing unless the value is set.

Patch-set: 3
Label: Code-Review=-1