Update patch set 2
Patch Set 2:

I don't feel too strongly, other than clean.sh is intended to reset state so the next run of stack.sh is 'mostly-fresh'.  I think the current setup is not going to  surprise a user and I can live with it.

But I also just realized that the flag file needs to be added to .gitignore.  ;)

Patch-set: 2