neutron: add NOVA_ALLOW_DUPLICATE_NETWORKS config option
Nova commit 322cc9336fe6f6fe9b3f0da33c6b26a3e5ea9b0c added the
neutron.allow_duplicate_networks config option in Juno and it defaults
to False. However, the option is now deprecated in Kilo with commit
4306d9190f49e7fadf88669d18effedabc880d3b and will be removed in Liberty.

To test it in the gate with Tempest, we need to be able to set it to
True in devstack-gate and update tempest.conf.

We can remove the change to lib/neutron when the option is removed from
Nova in Liberty, and we can remove the change to lib/tempest when it's
removed from Tempest after juno-eol.

Depends-On: I05f81d86cde249c23be06d5804fadbf40fc4a7f3

Change-Id: Ifd075420f57c9b60746f4a6af6520c0ef04800db