Update patch set 1
Patch Set 1:

We could also easily probe for BASH_VERSION ... we've always said it has to run under bash but never *explicitly* enforced it. Happy to make that change if we like.

Only thing is that the usual override mechanisms might not work ... we want to check this very very early before we've sourced any of our helper files. so adding FORCE_BASH=yes to local conf might be too late

Patch-set: 1