Update patch set 15
Patch Set 15: Code-Review+2

Should this be with the plugin or devstack?  

It's here, it's written, it's been reviewed so I think commit it.  If someone feels strongly enough to move it, we can just link to it.

In general; for existing large projects that already have a documentation structure setup, then it's probably logical for them to integrate their plugin documentation to that.  But if the plugin has to setup its own sphinx build and jobs etc just for some html pages, integrating here is probably better, IMO

Patch-set: 15
Label: Code-Review=+2