neutron: make l3/vpn agent use config-dir
Now that advanced services are split from the core tree, there are no
guarantees that any auxiliary configuration files are installed on the
system.

In devstack, it's handled by dynamically generating list of CLI
arguments based on which services are enabled in local.conf. This
approach is not applicable for distributions though where those
arguments are usually hardcoded in a service manifest.

Distributions will need to switch to using an approach that is more
dynamic. oslo.config provides us with an option to replace --config-file
arguments with --config-dir, that would load all *.conf files from
there. That would allow distributions to keep the same arguments no
matter which services are installed on a system.

This patch applies the same approach to devstack installations since
it's expected to be the way to go forward.

In this patch, I'm assuming that VPN agent cannot be used on the same
node as L3 agent. If that's not the case, we may need to maintain two
separate conf-dirs.

Note: we're leaving actual configuration files in their old paths, we
only create symbolic links to point to them from the configuration
directory.

Closes-Bug: #1431292
Depends-On: I072f35100737f4f139f647b508a22b45d40f13fc
Change-Id: I84fc20c4b6f58c2ae75bb1be226dcd783f81832e