Support DATABASE_MAX_CONNECTIONS for PostgreSQL backend
Use value of DATABASE_MAX_CONNECTIONS if it is non-zero instead of
MAX_DB_CONNECTIONS for PostgreSQL. This way there should be no
regressions for anybody using the setting.

In the future the MAX_DB_CONNECTIONS can be removed.

200 remains as default for PostgreSQL. This can be moved to stackrc in
case we agreed to use the same default setting for all database
backends (currently PostgreSQL and MySQL).

Change-Id: I61ed0a66cea836d52e37d5eeae97fa413fcb7783