Update patch set 4
Patch Set 4: Code-Review-1

Are we actually solving a problem here, or just adding variables for the sake of it?

something in the DATABASE_* namespace (like DATABASE_USER / DATABASE_PASSWORD) to my mind should be a totally generic thing across the global concept of the backing databases.

For example, if I set this to 200 for some reason and am using postgres, then switch my backend to mysql but don't update my local config -- does that imply i really want 200 connections for mysql as well?  IMO capacity planning for a db backend is not a generic tuning option and so shouldn't be exposed as such.

I think the confusion stems from the postgresql variable being badly named.  It should have been prefixed with a POSTGRES_* namespace identifier so it doesn't look like a global.  Mysql can have it's own -- so far nobody seems to have needed it.

Patch-set: 4
Reviewer: Gerrit User 7118 <7118@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1