Update patch set 35
Patch Set 35: Code-Review-1

My suggestion (which is not original, it has been raised by devstack core reviewers in this and other patches) is to 

1)start by documenting the known issues, so that users who work behind firewalls (myself included) can know and understand what variables are affected by proxies.  This documentation could provide examples and use cases about what to do in various firewall situations.

2) The next step would be to propose a comprehensive facility in devstack that could manage these environment settings, but rather than doing this automatically based on trying to guess programmatically what to do in a particular environment.  I hate to suggest a time-consuming and heavyweight process, but in this case I feel that it would be appropriate to write up a blueprint and discuss this entire problem at a more comprehensive level.

For example, perhaps it would be better to provide a flag in the config or command line to explicitly turn on "behind firewall" mode.  I do not like the idea of automatically altering the environment without me being aware of what is happening.

This work would need to address the devstack plugins that are now being required for all substantial new devstack work (see, for example, neutron-lbaas).

3) Work with other interested parties so design a robust, comprehensive solution rather than various individual patches that will not do the right thing in all cases.

https://review.openstack.org/#/c/153397/
https://review.openstack.org/#/c/154536/

Patch-set: 35