Use global_filter to avoid scan of missing devs
A while back I added an lvm.conf file with a device filter setting
to try and clean up the LVM hangs in the gate:
  (commit 0b9e76f280208b5b5ad54bb6fbc4133e63037286)

It turns out this was only part of the problem, the other problem
is that on an LVS/VGS command LVM will always poke at block devices
on the system to see if they belong to it.  There are a number of
situations in gate tests where /dev/sdX devices are created and
deleted and the result is that we hit situations where LVM tries
to open up devices to check them even if they've been removed.

This typically happens because udev hasn't updated yet, but the end
result is we have a blocking open call from LVM that takes approx
60 seconds to time out and fail.

The filter= only tells LVM to return devices from it's list that pass
the filter, but sadly it doesn't tell LVM that what it should or
shouldn't try to query via open calls. For that we need the global_filter.

This does in fact seem to eliminate our super long LVS/VGS command times
on several runs against my local setup.

This patch just adds an additional sed command to the cinder_backend lvm
setup script and makes the filter we already set global.

NOTE:
Another possible solution is to enable the lvmetad in our lvm.conf file.
This essentially tells LVM to store the device data when VG's or PV's are
added, this means we don't do the open at all which is likely a good
improvement that should be looked at.

There's some concern here though because it modifies the behavior of activate,
so I'd like to spend some time looking at that more closely and test things
out before considering using that instead of filters.

Change-Id: I6e202e01e627a161208b4c3745cfe5b8ef2ebe90
Closes-Bug: #1373513