Update patch set 1
Patch Set 1:

Well this is disappointing, the global as it's name implies only works in a global capacity.  In other words putting it in the local version of the lvm.conf doesn't seem to work, it must be in the base /etc/lvm/lvm.conf.

Unless my custom lvm conf isn't working at all... sigh

Patch-set: 1