Use global-requirements.gate to pre-install specific reqs (WIP)
This adds the ability to pre-install a specific set of dependencies
from global-requirements.gate.  The needed requirements are
calculated from each project's requirements.txt using pip-compile, and the
specific version are queried from global-requirements.gate (to be hosted in
the requirements repo). This results in a temporary requirements file that
is pip installed just prior to installing the project code.

Change-Id: I14703ee4ac53de049c9ab6c00a2580b3ce47ae4c