Update patch set 4
Patch Set 4:

Ian and everybody, thank you for reviewing this. I wish I could respond faster.


I'll explain why I thought it should be a single setting for both DB backends. In my mind, number of max connections a DB backend can have is something that should be generic in the context of using Devstack for tempest tests. If there are not enough connections for tempest workload with one backend, - and that was the reason for the change of PostgreSQL default with I2fcae8184a82e303103795a7bf57c723e27190c9 - there would probably be not enough connections if we switch the backend. It makes sense to have it in sync. E.g. if tempest concurrency ever increases, both might have to change.


The problem this patch was motivated by is inability to easily change the MySQL setting with Devstack-gate while trying to debug failures in our CI.


Having separate setting for MySQL does solve it, although adding more unnecessary variables. Is there ever a possibility for both backends to be used at the same time?

Patch-set: 4
Reviewer: Gerrit User 7979 <7979@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0