Update patch set 4
Patch Set 4:

@Mikhail : so I think we've converged to an understanding.

Is database connections a generic tuning parameter, or should each db backend provide it's own variable?

So if others think it's fine as a generic thing, i am happy with this change to implement that.  I'm still of the opinion each db should provide it's own name-spaced variable for such tuning ... but as I say, happy to be outvoted.

Patch-set: 4