Remove configure_cinder_backend_conf_lvm
Setting the local device filter like we were doing here
doesn't actually solve the problem.  We actually need to
set a global filter.

Submitted change: I5d5c48e188cbb9b4208096736807f082bce524e8

Once that lands this code won't be necessary at all and should
be removed.  We'll also remove the local conf usage in Cinder,
but currently that's harmless as it's designed to detect if
the local lvm.conf file exists in the cinder directory and
behave appropriately based on it's existence.

Change-Id: Iba701a8cb00ef7ce5acd675e8714ad471fb4cb37