Fix services' prereq packages not getting installed
The devstack commit 7ca90 changes the matching conditions for
how packages are determined to be installed for a given service.
Currently, if a service name is not seen in the list of files
to parse, then the service's package list file is added to the
list of files to parse. This breaks if the package file path
has the service name in it as the check will think the service's
package list file was already added.

Change-Id: Ie81352ebd5691afc6d0019f71d5b62370e8bb95f
Closes-Bug: #1434314