Update patch set 2
Patch Set 2: Code-Review-1

not everything needs to be in stackrc, so I think this should be in functions-common, otherwise things like grenade are going to revert. Put it in docs if we need it to be discoverable.

Patch-set: 2
Reviewer: Gerrit User 2750 <2750@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1