Update patch set 3
Patch Set 3: Code-Review-1

(2 comments)

Wiekus, please also add an entry to the MAINTAINERS.rst for OL.

Ian, FWIW I'd put this in the same category as SuSE, no testing is reported back to us and it often lags many changes because often we don't know what needs to be done.  With more and more differences between OL and RHEL we need to make that clear here too.

Patch-set: 3
Label: Code-Review=-1