Update patch set 20
Patch Set 20:

This was mainly an update to try and reduce duplication since the code on setup/cleanup were almost identical.  I also changed it to only move the primary IP to keep it simple, leaving the existing code to clean up the rest.

So I'll abstain from voting on this one.

Patch-set: 20
Label: Verified=0