Update patch set 1
Patch Set 1: Code-Review-1

I'm not seeing how this is consumed?  Is it via Ib48d652fba8dabe1e94203d44278b2bae56b356c

the comments about HOST_IP and how to access it would have been useful in the changeset to motivate the change

this can't be found in the pre-hook, but a post-hook should be able to see and source .stackenv which has this in it.  can we leverage that?

Patch-set: 1
Reviewer: Gerrit User 7118 <7118@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1