Update patch set 9
Abandoned

I'm going to take this out of the devstack queue because it shouldn't be done here.

As far as I can see, oslo-incubator does not have the correct requirement on python-memcached when it imports that package.  we should fix that and bring it in so nova has the correct dependencies.  See I14971ecbeb45943d53c524aef52141f235b91060

Patch-set: 9
Status: abandoned