Remove unused IPv6 setup from DevStack neutron script
With the implementation of dual-stack gateway support
as part of the Neutron multiple-ipv6-prefixes
blueprint, some of the code in the IPv6 setup in the
DevStack neutron script can be removed. This code
had been added temporarily, with a TODO note
indicating that this code should be removed when
the Neutron L3 agent is modified to support
dual-stack.

This change should not be merged until after
the changes for:
https://review.openstack.org/#/c/161085/
(and changes that this depends upon) have been
merged.

Change-Id: I85fe68782bc54f28f3e14aa4a1d042cb15959dac
Depends-On: I7d4e8194815e626f1cfa267f77a3f2475fdfa3d1
Depends-On: I9395834f673038dc23b25eaeefe14895fe154e0e
Partially-implements: blueprint multiple-ipv6-prefixes