Update patch set 1
Patch Set 1: Code-Review-1

I like having the option setup, but I don't think we're ready to make it the default until it's the default in Cinder.  I'm happy to work on this as a first change in Liberty, but I don't think we should mess with it for Kilo.

Patch-set: 1
Reviewer: Gerrit User 2243 <2243@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1