Prepare devstack for Ironic cleaning testing
This patch changes the two config options required for Ironic to
successfully test cleaning in devstack.

First, we disable erase_devices clean step. Erase devices in VMs
ends up running shred on the drives for the agent driver, which is
incredibly slow and completely unneeded in devstack.

Additionally, we allow Ironic more time to complete the unprovision
after the nova instance is deleted. This time is spend in the CLEANING
state to clean up after the node.

This is related to the Ironic blueprint "implement-cleaning-states".

Change-Id: I77081165a80491da3e66d8a4554b6d71fc3d9353