[WIP] Switch the default to Linuxbridge
This patch was created in response to a discussion regarding the
migration from nova-network to neutron [1]. It was argued that
the default of neutron using ovs would make the setup too complicated
for some operators, so we are looking here at the option of running
with linuxbridge as the default instead.

Note that this is still WIP, the intention of the patch is to find out
which things are breaking in the Jenkins tests and how to solve them.

[1] http://lists.openstack.org/pipermail/openstack-dev/2015-March/060029.html

Co-Authored-By: Jens Rosenboom <j.rosenboom@x-ion.de>

Change-Id: Ie7436d68b6921cd0db9eb28b55bc74583345c04f