Begin new lib/neutron
* minor tweaks to lib/template
* add new lib/neutron
* begin fleshing out the functions requried for api/agent service only
* add dispatch functions to handle the function name collisions between lib/neutron
  and lib/neutron-legacy

Usage of the new library is by setting the following in ENABLED_SERVICES

* neutron-api
* neutron-l3
* neutron-agent
* neutron-dhcp
* neutron-metadata-agent

For now, the new neutron library supports just the ML2 plugin. Ideally
different neutron plugins would become DevStack plugins. Work needs to
be done in this commit to make that possible, since there are lots of
configuration options that need to be set prior to the start of
neutron-api and other services.

Co-Authored-By: Sean M. Collins <sean@coreitpro.com>
Co-Authored-By: Hirofumi Ichihara <ichihara.hirofumi@lab.ntt.co.jp>
Depends-On: I166dc43372eef0243d53e3846dbc8e7a216317d2
Change-Id: I31b6362c6d9992f425f2dedbbeff2568390a93da