Assurance status of rabbitmq is running
I changed it so that rabbitmq restart in case of ubuntu as well as fedora , suse.

Current devstck don't restart rabbitmq in case of ubuntu. Becasue
rabbitmq is running at default.
But this approach have the following bug.
If rabbitmq is already installed and not running ,  stack.sh will
fail.
So I change it so that rabbitmq restart in case of ubuntu.

Closes-bug: #1030798
Change-Id: Ie45446d3817b2f15631f03b2af84749fe936c67b