Restart rabbitmq if rabbitmq-server isn’t running
In case of ubuntu.
Rabbitmq-server is running at default , if using apt-get installer.
So currently we don’t restart rabbitmq-server .
But this approach have the following bug.
If rabbitmq-server is already installed and not running ,  stack.sh will
fail.
This commit change it so that rabbitmq-server is restarted  if rabbitmq
is not running.

Change-Id: Ie45446d3817b2f15631f03b2af84749fe936c67b