Update patch set 7
Patch Set 7: Code-Review-1

(1 comment)

Beyond doing the XML work in Python like we talked about in IRC, I think the only other real change this needs is doing it as part of configure_libvirt() in lib/nova_plugins/functions-libvirt.  That is where the rest of the system-level libvirt bits are set.  If you need to change the restart_service at the bottom to a stop_service()/start_service() pair, go ahead.

Patch-set: 7
Label: Code-Review=-1