Catalog Index Service - glance devstack
Implements: blueprint catalog-index-service

The changes to lib/glance incorporate the new g-search service.
The g-search service is optional.
To enable it add the following line to devstack/local.conf:
enable_service g-search

In addition to deploying g-search, the changes will also populate a
search type of keystone service and adds in appropriate endpoints.

Change-Id: I0272d56bc2e50e8174db78bd449f65f60f7f4000