Update patch set 4
Patch Set 4:

So that works better now, only one failure remaining and I'm not even sure that it is related: http://paste.openstack.org/show/197752/

I won't have time to continue on this before the weedend, so if s/o wants to make a proper patch from this, please go ahead. Maybe after this works, one can also try to go back to the vxlan version, if that is preferred as a smaller change for the first step.

Patch-set: 4
Label: Verified=0