Update patch set 1
Patch Set 1: Code-Review+1

The default for those used to be localhost...that's fixed??  In Kilo I'm guessing?  Woot!

I'm only +1 for now as I have not yet built a new test environment and really want to make sure OSC will be as happy about this as I think it will...

Patch-set: 1