Update patch set 1
Patch Set 1:

Melwitt: interesting thought.  I don't want to give the impression that everything works with networking so we need to find the right balance there.  But anything that uses nova-manage to create networks should work as long as operators run the command on parent and child databases, which we should document.

Patch-set: 1