sleep less during screen_it
the sleep 3 in screen_it was added to make devstack pass in the gate
with exceptionally slow test cloud nodes. In the gate we now bypass
the screen path entirely. However the sleep 3 remains and can add a
couple minutes delay into local development runs.

Tune this down to 0.1s, which should be fine for the local case. It
can also be overridden if people see lots of fails in their local
environment.

Change-Id: I0e6476176fc8589efc4e40e78c2231f704d14e45