Fix readme for multi-node setup
The readme stated that a compute node in a multi-node setup requires the
glance-api service to be enabled. But actually the glance-api service is only
required on the controller node where glance-registry is running.
Running the glance-api service on a node without glance-registry will even lead
to a failure of glance-api, as the glance cache directory will not be created
without enabling glance-registry.

Change-Id: Ie92533f3333f3fe3e2d747762e60f2f42a233e79
Closes-bug: #1441198