Update patch set 2
Patch Set 2: Code-Review-1

i'd just like to confirm we don't have some other bug before we land this... I'm guessing this started breaking when the service reduction change went in.

 [iwienand@fedora19 devstack]$ git grep libxml2
 files/debs/ceilometer-collector:libxml2-dev
 files/debs/devlibs:libxml2-dev  # lxml
 files/debs/glance:libxml2-dev
 files/debs/nova:python-libxml2

so nova should have brought in python-libxml2, which should be 2.9.1 on a trusty host which is greater than what you need there

 Downloading/unpacking lxml>=2.3 (from -r /opt/stack/tempest/requirements.txt (line 7))

is nova not installed on this host?  

[1] http://packages.ubuntu.com/trusty/python-libxml2

Patch-set: 2