Skip cert related things in create_userrc.sh when n-crt absent
169709 removed n-crt from the default services, but missed the fact that
create_userrc.sh is doing lots of cert things by default, some of which
is dependent on that service being active.  To avoid errors and long RPC
timeouts trying to run nova x509-* commands, this adds a new flag to skip
cert stuff that is passed when the service is missing.

Change-Id: Ib5023f9dd681fbd363510a93166ec20bda7dc923
Closes-bug: #1441348