Update patch set 2
Patch Set 2: Code-Review+2

so it's libxslt that's actually missing.  it is still not clear from the partial log how things ended up like this [1].

the long term intent is I guess to build this as a wheel and use that.  

in the mean time, might as well do it like this -- as long as nova is including the packaged version that's going to override most everything in a single-node version anyway; i.e. installing the dev packages is just wasted time and space in that case so they don't need to be in "general"

[1] https://launchpadlibrarian.net/150255293/log

Patch-set: 2
Reviewer: Gerrit User 7118 <7118@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+2