Update patch set 1
Patch Set 1:

which is why it's overrideable. As dtroyer said in the juno, we lived with this as 0.5 for *a long time* until we ended up on some really slow cloud boxes with noisy neighbors, when I eventually crept it up to 3 seconds. In doing so we added minutes of delay to the common case of running devstack in a local vm or on a real box. I'm fine if we want to default back to 0.5 instead of 0.1, but I'm not interested in adding complexity here we have to debug, or in keeping the long delays in place.

Patch-set: 1