Update patch set 11
Patch Set 11:

O.k., back to using vlan instead of vxlan, all checks are passing except for the ironic issue I mentioned earlier and some issue in check-tempest-dsvm-f21 which seems unrelated.

So now we can start discussing how the defaults should look like in order for a user to be able to create a working devstack from this without having to set up anything special.

IMO running on eth0 as default would be fine, for the standard setup running inside an instance and only needing local connectivity that should just work, as the successful tests show.

Patch-set: 11
Label: Verified=0