Specify network UUID for network create with cells
When the n-cell service is enabled, we create networks in both the
API cell and the child cell. Recent changes to tempest have tests
querying networks from the API and passing them for a server create.
In order for this to work in cells, the UUIDs for the network in
the API cell and the child cell must match, else the network won't
be found in the child.

This change adds the --uuid option to the nova-manage network create
command for cells only.

Related-Bug: #1441931

Depends-On: Ib29e632b09905f557a7a6910d58207ed91cdc047

Change-Id: Ib5933b1405c0761ff727e04cda0c502a826c8eaf