[WIP] Add a peak memory tracker to dstat
We can see memory usage during the run with dstat but we have no
insight into what is consuming the memory.

This adds a tracker to dstat that records a snapshot of peak memory
usage during a run.

The intent is to add logic into the verify-pipeline jobs to use this
report and send statistics on peak memory usage to statsd [1].  We can
then build a picture of memory-usage growth over time.  This should
allow better insight into an issues such as that introduced by
Idf3a3a914b54779172776822710b3e52e751b1d1 where memory-usage jumped
dramatically after switching to pip versions of libraries.

dstat output on existing runs shows that available memory usage is at
peak times down to under 150MiB on our 8GiB CI hosts ... this is not a
lot of head-room.  So tracking memory growth is going to be an
important part of the future.

[1] http://graphite.openstack.org/

Change-Id: I4b0a8f382dcaa09331987ab84a68546ec29cbc18