Update patch set 1
Patch Set 1:

I am not sure glance-api and glance-registry can sit on different nodes:

init_glance is the function that (re)creates the SQL database. Currently init_glance is called when g-reg is enabled. So if you have glance-api and glance-registry on different nodes and the node with g-api starts first, you will run into issue because the SQL DB will not be there.

Patch-set: 1
Label: Verified=0