Set live_migrate_paused_instances=True in tempest.conf
Live migration of paused instances is a new Nova feature in Kilo, and will not be backported.
The compute_feature_enabled.live_migrate_paused_instances flag defaults to False for
this reason, but can be set to True here. The tempest config option and this change can
both be removed at Juno-EOL.

Change-Id: I3a83e43d252b88c234438a224e2fbebc0a81eaff
Depends-On: I5c6fd3de7ea45d1851bb40037c64ad7fb5e6dc48
Related-Bug: #1305062