Update patch set 5
Patch Set 5:

Chmouel: I think we understand the benefits, and I think we're all for it.  What would be good is a short description of how to set it up.  Sean mentioned the command, but not where you need to execute it.  I assume it's to be included in 'local.conf' under the [[local|localrc]] section, but I hate assuming ;)  Let me know if that assumption is correct and I'll start testing it out. Thx!

Patch-set: 5