Update patch set 1
Patch Set 1:

What I tried to do was to install a multi-node setup as explained in the readme.
In this setup I have one controller node, which runs glance-registry and glance-api, and a compute node which only runs glance-api. The glance-api on the compute node uses the glance-registry on the controller node (at least thats how I understand it).
It is clear that I have to start the glance-registry on the controller node before I can start the glance-api on the compute node, but even after glance-registry has started, glance-api will not start on the compute node because the glance cache directory is not there (it only gets created on the controller node when glance-registry starts).
So what I try to do in this patch is to create the glance cache directory when glance-api is enabled but glance-registry is not.

Patch-set: 1
Reviewer: Gerrit User 13998 <13998@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0