osc timing utility
This small script parses a devstack log and accounts for the time
spent running the openstack client. In a current run this is 4.5
minutes. While many of these calls are rest calls, and that's where
the time is, the fact that 1/3 of our setup time is inside OSC seems
problematic.

Change-Id: Idd5bb9b497bd791d70e6df270c196efa011617dc