Fix non group-set $command for _run_process
The "sg" command takes a quoted string and runs it in a shell, while
the non-group path isn't doing this.  This means the function behaves
differently depending on if you specify a group or not -- "cd /blah;
cmd" works with a group set but not otherwise.  Bring into line by
running with bash -c

Change-Id: Ic854ebd029aef93bfee620d92d244f2c78766d18