Update patch set 1
Patch Set 1:

Ok, I just did some testing and it looks like a compute node does not even need the glance-api service to be enabled.

This makes my fix rather irrelevant, as it seems like this is just a problem in the documentation of the multi-node setup.

I will just overwrite my patch with a fix for the readme.

Patch-set: 1