Actually run function tests
run_tests.sh runs all scripts tests/test_* as for the upstream
unit-test job, but a lot of tests are in "functions.sh" which isn't
picked up.  The existing "test_functions.sh" is only testing
true/false function -- move that into its own sub-test (it also uses
unittest.sh functions; easier to leave it alone) and move
"functions.sh" to "test_functions.sh" so that it is run.

Also update comment in run_tests.sh, which seems to be an old
reference to when less testing was done.

Change-Id: I3f6042f1ad324fc03d45c16e8dbf95b751d62e5e