Update patch set 6
Patch Set 6: Code-Review+2

Thanks Wayne.  It's working now, but there may be a sticking point when glance-search is added in Grenade.  init_glance() is not called after the upgrades are done, if elasticsearch is stopped at that point it will not be re-started.

We may need to look a the semantics of the pkg/* calls again.

Patch-set: 6
Reviewer: Gerrit User 970 <970@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+2